+++ packages/react/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @micro-stacks/react
 
+## 1.0.0-beta.0
+
+### Major Changes
+
+- This is the first major version release of the new micro-stacks client and react libraries!
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @micro-stacks/client@1.0.0-beta.0
+
 ## 0.2.1
 
 ### Patch Changes